Update meters-update-display-optional<-development
Including reconciling package-lock.json, 0.3.0 to 0.4.0 migration, and
making the baseline tests use the right Meter constructor.
