+++ .travis.yml
@@ -4,7 +4,7 @@
 # * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 # *
 
-dist: trusty
+dist: xenial
 sudo: required
 
 language: node_js
