+++ CONTRIBUTING.md
@@ -1,8 +1,17 @@
 # Contributing to OED
 
+We're thrilled that you're interested in contributing to the OED project. We
+welcome all open source contributions, whether bug reports, patches, or anything
+else.
+
+Check out the [USAGE.md](USAGE.md) file to learn how you can set up OED on your
+own system.
+
 ## Legal Stuff
 
-Please sign our [CLA](https://goo.gl/forms/nR9MtVHUOqYn8WbP2) before contributing to OED.
+Please sign our [CLA](https://goo.gl/forms/nR9MtVHUOqYn8WbP2) before
+contributing code to OED. It is not necessary to sign the CLA in order to
+contribute by filing or discussing issues.
 
 ## Development Workflow
