+++ src/server/models/Migration.js
@@ -72,7 +72,7 @@ class Migration {
 		if (rows.length > 0) {
 			return rows.map(row => new Migration(row.id, row.from_version, row.to_version, row.update_time));
 		} else {
-			throw new Error('There is no item in migration table or table does not exists');
+			throw new Error('There is nothing in the migration table.');
 		}
 	}
 }
